test(tooling): share the client HMR probe environment
